Basic miniplayer and tabbar functionality implemented
