+++ Tempo/Profile/UsersViewController.swift
@@ -35,7 +35,7 @@ class UsersViewController: UIViewController, UITableViewDelegate, UITableViewDat
 		extendedLayoutIncludesOpaqueBars = true
 		definesPresentationContext = true
 		
-		tableView = UITableView(frame: CGRect(x: 0, y: 0, width: UIScreen.main.bounds.width, height: UIScreen.main.bounds.height - playerCellHeight), style: .plain)
+		tableView = UITableView(frame: CGRect(x: 0, y: 0, width: UIScreen.main.bounds.width, height: UIScreen.main.bounds.height - tabBarHeight), style: .plain)
 		tableView.delegate = self
 		tableView.dataSource = self
 		tableView.rowHeight = 101
@@ -74,7 +74,6 @@ class UsersViewController: UIViewController, UITableViewDelegate, UITableViewDat
 		if displayType == .Users {
 			title = "People"
 			
-			addHamburgerMenu()
 			populateSuggestions()
 		} else {
 			showActivityIndicator()
